backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
starlette==0.36.3
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
urllib3==2.1.0
openai==1.12.0
redis==5.0.1
tenacity==8.2.3
celery==5.3.6
flower==2.0.1 